Add default values to properties
In case the active property isn't set it should default to true. The year property needs to be a number to be stored in the db but arrives as a string. Convert it to a number by adding a plus sign in front of the value when it's assigned to a variable.